requirements.txt
django
djangorestframework
drf-extensions
coreapi
markdown
pygments
mlswarm-infrastructure
